refactor(OGSpy): :fire: Moving ogame.php content to dedicated files to have better understanding of calculation
