+++ tests/unit/IpEncodingTest.php
@@ -10,7 +10,8 @@ final class IpEncodingTest extends TestCase
      */
     public function testEncodeIpIPv4()
     {
-        require_once 'includes/functions.php';
+    if (!defined('IN_SPYOGAME')) define('IN_SPYOGAME', true);
+    require_once 'includes/functions.php';
         $ip = '192.168.0.1';
         $expected = 'c0a80001';
         $result = encode_ip($ip);
@@ -19,7 +20,8 @@ final class IpEncodingTest extends TestCase
 
     public function testEncodeIpIPv6()
     {
-        require_once 'includes/functions.php';
+    if (!defined('IN_SPYOGAME')) define('IN_SPYOGAME', true);
+    require_once 'includes/functions.php';
         $ip = '2001:0db8:85a3:0000:0000:8a2e:0370:7334';
         $expected = '20010db885a3000000008a2e03707334';
         $result = encode_ip($ip);
@@ -29,7 +31,8 @@ final class IpEncodingTest extends TestCase
 
     public function testDecodeIpReturnsIPv4FormatForIPv4EncodedInput()
     {
-        require_once 'includes/functions.php';
+    if (!defined('IN_SPYOGAME')) define('IN_SPYOGAME', true);
+    require_once 'includes/functions.php';
         $hexIp = 'c0a80001'; // IPv4 address encoded in hex format
         
         $decodedIp = decode_ip($hexIp);
@@ -39,7 +42,8 @@ final class IpEncodingTest extends TestCase
     
     public function testDecodeIpReturnsIPv6FormatForIPv6EncodedInput()
     {
-        require_once 'includes/functions.php';
+    if (!defined('IN_SPYOGAME')) define('IN_SPYOGAME', true);
+    require_once 'includes/functions.php';
         $hexIp = '20010db885a3000000008a2e03707334'; // IPv6 address encoded in hex format
         
         $decodedIp = decode_ip($hexIp);
